Standalone mostly reincarnated.
