+++ RecoTracker/MkFit/plugins/MkFitProducer.cc
@@ -20,7 +20,7 @@
 // mkFit includes
 #include "RecoTracker/MkFitCore/interface/ConfigWrapper.h"
 #include "RecoTracker/MkFitCMS/interface/LayerNumberConverter.h"
-#include "RecoTracker/MkFitCMS/interface/buildtestMPlex.h"
+#include "RecoTracker/MkFitCMS/interface/runFunctions.h"
 #include "RecoTracker/MkFitCore/interface/IterationConfig.h"
 #include "RecoTracker/MkFitCore/interface/MkBuilderWrapper.h"
