+++ RecoTracker/MkFitCMS/interface/runFunctions.h
@@ -1,5 +1,5 @@
-#ifndef RecoTracker_MkFitCMS_interface_buildtestMPlex_h
-#define RecoTracker_MkFitCMS_interface_buildtestMPlex_h
+#ifndef RecoTracker_MkFitCMS_interface_runFunctions_h
+#define RecoTracker_MkFitCMS_interface_runFunctions_h
 
 #include "RecoTracker/MkFitCore/interface/Track.h"
