+++ RecoTracker/MkFitCMS/standalone/buildtestMPlex.cc
@@ -1,7 +1,7 @@
-#include "RecoTracker/MkFitCMS/interface/buildtestMPlex.h"
 #include "RecoTracker/MkFitCore/src/Matrix.h"
 #include "RecoTracker/MkFitCore/interface/MkBuilder.h"
 #include "RecoTracker/MkFitCMS/interface/MkStdSeqs.h"
+#include "RecoTracker/MkFitCMS/standalone/buildtestMPlex.h"
 
 #include "tbb/parallel_for.h"
 
@@ -92,9 +92,12 @@ namespace mkfit {
 
     builder.begin_event(&job, &ev, __func__);
 
+    // CCCC
+    /*
     if (Config::sim_val_for_cmssw) {
       builder.root_val_dumb_cmssw();
     }
+    */
     builder.end_event();
   }
 
@@ -132,7 +135,7 @@ namespace mkfit {
 
     builder.begin_event(&job, &ev, __func__);
 
-    builder.PrepareSeeds();
+    // CCCC builder.PrepareSeeds();
 
     // EventOfCandidates event_of_cands;
     builder.find_tracks_load_seeds_BH(ev.seedTracks_);
@@ -160,7 +163,7 @@ namespace mkfit {
     if (Config::quality_val || Config::sim_val || Config::cmssw_val) {
       //Mark tracks as duplicates; if within CMSSW, remove duplicate tracks before backward fit
       if (Config::removeDuplicates) {
-        StdSeq::find_duplicates(ev.candidateTracks_);
+        // CCCC StdSeq::find_duplicates(ev.candidateTracks_);
       }
     }
 
@@ -170,11 +173,14 @@ namespace mkfit {
       ev.fitTracks_ = builder.ref_tracks();
     }
 
+    // CCCC
+    /*
     if (Config::quality_val) {
       builder.quality_val();
     } else if (Config::sim_val || Config::cmssw_val) {
       builder.root_val();
     }
+    */
     builder.end_event();
 
     // ev.print_tracks(ev.candidateTracks_, true);
@@ -216,7 +222,7 @@ namespace mkfit {
 
     builder.begin_event(&job, &ev, __func__);
 
-    builder.PrepareSeeds();
+    // CCCC builder.PrepareSeeds();
 
     builder.find_tracks_load_seeds(ev.seedTracks_);
 
@@ -239,7 +245,7 @@ namespace mkfit {
     check_nan_n_silly_candidates(ev);
 
     // first store candidate tracks
-    builder.quality_store_tracks(ev.candidateTracks_);
+    // CCCC builder.quality_store_tracks(ev.candidateTracks_);
 
     // now do backwards fit... do we want to time this section?
     if (Config::backwardFit) {
@@ -252,14 +258,17 @@ namespace mkfit {
       check_nan_n_silly_bkfit(ev);
     }
 
-    StdSeq::handle_duplicates(&ev);
+    // CCCC StdSeq::handle_duplicates(&ev);
 
     // validation section
+    // CCCC
+    /*
     if (Config::quality_val) {
       builder.quality_val();
     } else if (Config::sim_val || Config::cmssw_val) {
       builder.root_val();
     }
+    */
 
     builder.end_event();
 
@@ -302,7 +311,7 @@ namespace mkfit {
 
     builder.begin_event(&job, &ev, __func__);
 
-    builder.PrepareSeeds();
+    // CCCC builder.PrepareSeeds();
 
     builder.find_tracks_load_seeds(ev.seedTracks_);
 
@@ -325,7 +334,7 @@ namespace mkfit {
     check_nan_n_silly_candidates(ev);
 
     // first store candidate tracks - needed for BH backward fit and root_validation
-    builder.quality_store_tracks(ev.candidateTracks_);
+    // CCCC builder.quality_store_tracks(ev.candidateTracks_);
 
     // now do backwards fit... do we want to time this section?
     if (Config::backwardFit) {
@@ -341,14 +350,17 @@ namespace mkfit {
       check_nan_n_silly_bkfit(ev);
     }
 
-    StdSeq::handle_duplicates(&ev);
+    // CCCC StdSeq::handle_duplicates(&ev);
 
     // validation section
+    // CCCC
+    /*
     if (Config::quality_val) {
       builder.quality_val();
     } else if (Config::sim_val || Config::cmssw_val) {
       builder.root_val();
     }
+    */
 
     builder.end_event();
 
@@ -527,11 +539,14 @@ namespace mkfit {
     MkJob job({Config::TrkInfo, Config::ItrInfo[0], eoh});
     builder.begin_event(&job, &ev, __func__);
 
+    // CCCC
+    /*
     if (validation_on) {
       builder.prep_simtracks();
       //swap for the cleaned seeds
       ev.seedTracks_.swap(seeds_used);
     }
+    */
 
     check_nan_n_silly_candidates(ev);
 
@@ -539,11 +554,14 @@ namespace mkfit {
       check_nan_n_silly_bkfit(ev);
 
     // validation section
+    // CCCC
+    /*
     if (Config::quality_val) {
       builder.quality_val();
     } else if (Config::sim_val || Config::cmssw_val) {
       builder.root_val();
     }
+    */
 
     // ev.print_tracks(ev.candidateTracks_, true);
 
@@ -554,107 +572,6 @@ namespace mkfit {
     builder.release_memory();
 
     return timevec;
-  }
-
-  //==============================================================================
-  // run_OneIteration
-  //
-  // One-stop function for running track building from CMSSW.
-  //==============================================================================
-
-  struct IterationMaskIfcCmssw : public IterationMaskIfcBase {
-    const TrackerInfo &m_trk_info;
-    const std::vector<const std::vector<bool> *> &m_mask_vector;
-
-    IterationMaskIfcCmssw(const TrackerInfo &ti, const std::vector<const std::vector<bool> *> &maskvec)
-        : m_trk_info(ti), m_mask_vector(maskvec) {}
-
-    const std::vector<bool> *get_mask_for_layer(int layer) const {
-      return m_trk_info.m_layers[layer].is_pix_lyr() ? m_mask_vector[0] : m_mask_vector[1];
-    }
-  };
-
-  void run_OneIteration(const TrackerInfo &trackerInfo,
-                        const IterationConfig &itconf,
-                        const EventOfHits &eoh,
-                        const std::vector<const std::vector<bool> *> &hit_masks,
-                        MkBuilder &builder,
-                        TrackVec &seeds,
-                        TrackVec &out_tracks,
-                        bool do_seed_clean,
-                        bool do_backward_fit,
-                        bool do_remove_duplicates) {
-    IterationMaskIfcCmssw it_mask_ifc(trackerInfo, hit_masks);
-
-    MkJob job({trackerInfo, itconf, eoh, &it_mask_ifc});
-
-    builder.begin_event(&job, nullptr, __func__);
-
-    if (do_seed_clean) {
-      // Seed cleaning not done on pixelLess / tobTec iters
-      if (itconf.m_requires_dupclean_tight)
-        StdSeq::clean_cms_seedtracks_iter(&seeds, itconf, eoh.m_beam_spot);
-    }
-
-    // Check nans in seeds -- this should not be needed when Slava fixes
-    // the track parameter coordinate transformation.
-    builder.seed_post_cleaning(seeds);
-
-    if (itconf.m_requires_seed_hit_sorting) {
-      for (auto &s : seeds)
-        s.sortHitsByLayer();  // sort seed hits for the matched hits (I hope it works here)
-    }
-
-    builder.find_tracks_load_seeds(seeds);
-
-    builder.FindTracksCloneEngine();
-
-    if (itconf.m_requires_quality_filter && itconf.m_track_algorithm != 7) {
-      if (itconf.m_track_algorithm == 6) {
-        builder.filter_comb_cands([&](const TrackCand &t) { return StdSeq::qfilter_n_hits_pixseed(t, 3); });
-      } else if (itconf.m_track_algorithm == 9) {
-        builder.filter_comb_cands(
-            [&](const TrackCand &t) { return StdSeq::qfilter_pixelLessFwd(t, eoh.m_beam_spot, trackerInfo); });
-      } else {
-        builder.filter_comb_cands(
-            [&](const TrackCand &t) { return StdSeq::qfilter_n_hits(t, itconf.m_params.minHitsQF); });
-      }
-    }
-
-    if (do_backward_fit) {
-      if (itconf.m_backward_search) {
-        builder.CompactifyHitStorageForBestCand(itconf.m_backward_drop_seed_hits, itconf.m_backward_fit_min_hits);
-      }
-
-      builder.BackwardFit();
-
-      if (itconf.m_backward_search) {
-        builder.BeginBkwSearch();
-        builder.FindTracksCloneEngine(SteeringParams::IT_BkwSearch);
-        builder.EndBkwSearch();
-      }
-
-      if (itconf.m_requires_quality_filter && (itconf.m_track_algorithm == 7 || itconf.m_track_algorithm == 9)) {
-        if (itconf.m_track_algorithm == 7) {
-          builder.filter_comb_cands(
-              [&](const TrackCand &t) { return StdSeq::qfilter_n_layers(t, eoh.m_beam_spot, trackerInfo); });
-        } else if (itconf.m_track_algorithm == 9) {
-          builder.filter_comb_cands(
-              [&](const TrackCand &t) { return StdSeq::qfilter_pixelLessBkwd(t, eoh.m_beam_spot, trackerInfo); });
-        }
-      }
-    }
-
-    builder.filter_comb_cands([&](const TrackCand &t) { return StdSeq::qfilter_nan_n_silly(t); });
-
-    builder.export_best_comb_cands(out_tracks, true);
-
-    if (do_remove_duplicates) {
-      StdSeq::find_and_remove_duplicates(out_tracks, itconf);
-    }
-
-    builder.end_event();
-    builder.release_memory();
   }
 
 }  // end namespace mkfit
