+++ RecoTracker/MkFitCore/standalone/Makefile.config
@@ -106,7 +106,7 @@ WITH_TBB := 1
 # Derived settings
 ################################################################
 
-CPPFLAGS := -I. ${USE_INTRINSICS}
+CPPFLAGS := ${USE_INTRINSICS}
 CXXFLAGS := -fPIC ${OPT} ${OSX_CXXFLAGS}
 
 LDFLAGS_HOST := 
@@ -190,6 +190,3 @@ endif
 CPPFLAGS += ${CPPUSERFLAGS}
 CXXFLAGS += ${CXXUSERFLAGS}
 LDFLAGS  += ${LDUSERFLAGS}
-
-%.d: %.cc
-	${MAKEDEPEND} -o $@ $<
